+++ gen/templates/provider.go
@@ -222,6 +222,7 @@ func (p *SdwanProvider) Resources(ctx context.Context) []func() resource.Resourc
 		{{- end}}
 		NewCLIDeviceTemplateResource,
 		NewFeatureDeviceTemplateResource,
+		NewAttachFeatureDeviceTemplateResource,
 	}
 }