+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1 (unreleased)
+
+- Fix device template configuration of sub-templates
+- Fix mandatory feature template attributes when using device-specific variables
+- Add `sdwan_attach_feature_device_template` resource
+
 ## 0.1.0
 
 - Initial release